FIX ProteinDataset returned SeqRec instead of Seq; TST additional tests and some refactoring; MNT removed fileshare https work-around
