+++ deepnog/tests/data/test_inference_short_wrong_column_names.csv
@@ -1,3 +1,3 @@
-,string_id,eggnog_id
+,string_id,class_id
 0,0,COG0443
 0,1,COG0443
